RAG/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pymongo==4.7.2
openai==1.30.1
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
duckduckgo_search==8.1.1